FEM: unit tests, move object tests in separate test module and class
