+++ src/Mod/Fem/CMakeLists.txt
@@ -118,6 +118,7 @@ SET(FemTests_SRCS
     femtest/testccxtools.py
     femtest/testcommon.py
     femtest/testmesh.py
+    femtest/testobject.py
     femtest/testsolverframework.py
     femtest/testtools.py
 )
