+++ src/Mod/Fem/TestFem.py
@@ -28,6 +28,7 @@
 # the order should be as follows:
 # common-, object-, mesh-, inout-, ccxtools-, solverframworktests
 from femtest.testcommon import FemCommon
+from femtest.testobject import FemObject
 from femtest.testmesh import FemMeshTest
 from femtest.testccxtools import FemCcxAnalysisTest
 from femtest.testsolverframework import SolverFrameWorkTest
